refactor: move storex to store
