+++ src/componentsX/VolumeRendering.vue
@@ -55,9 +55,9 @@ import vtkImageData from '@kitware/vtk.js/Common/DataModel/ImageData';
 import { Vector3 } from '@kitware/vtk.js/types';
 import { useResizeObserver } from '../composables/useResizeObserver';
 import { manageVTKSubscription } from '../composables/manageVTKSubscription';
-import { useView3DStore } from '../storex/views-3D';
+import { useView3DStore } from '../store/views-3D';
 import { useProxyManager } from '../composables/proxyManager';
-import { useDatasetStore } from '../storex/datasets';
+import { useDatasetStore } from '../store/datasets';
 import { createVolumeThumbnailer } from '../utils/volume-thumbnailer';
 import { useCurrentImage } from '../composables/useCurrentImage';
 import { useCameraOrientation } from '../composables/useCameraOrientation';
