+++ src/utils/manipulators.ts
@@ -1,7 +1,7 @@
 import { Vector3 } from '@kitware/vtk.js/types';
 import vtkPlaneManipulator from '@kitware/vtk.js/Widgets/Manipulators/PlaneManipulator';
 import { vec3 } from 'gl-matrix';
-import { ImageMetadata } from '../storex/datasets-images';
+import { ImageMetadata } from '../store/datasets-images';
 import { getLPSAxisFromDir, LPSAxisDir } from './lps';
 
 export function updatePlaneManipulatorFor2DView(
